# docs/swagger.yaml
definitions:
  project.Request:
    properties:
      description:
        example: text generator
        type: string
        x-order: "2"
      end_at:
        example: "2006-01-02"
        type: string
        x-order: "4"
      manager_id:
        example: id
        type: string
        x-order: "5"
      started_at:
        example: "2006-01-02"
        type: string
        x-order: "3"
      title:
        example: ai startup
        type: string
        x-order: "1"
    type: object
  response.Object:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  task.Request:
    properties:
      completion_at:
        example: "2006-01-02"
        type: string
        x-order: "7"
      description:
        example: add new feature to the project
        type: string
        x-order: "2"
      priority:
        example: high
        type: string
        x-order: "3"
      project_id:
        example: id
        type: string
        x-order: "6"
      responsible_id:
        example: id
        type: string
        x-order: "5"
      status:
        example: in_progress
        type: string
        x-order: "4"
      title:
        example: add new feature
        type: string
        x-order: "1"
    type: object
  user.Request:
    properties:
      email:
        example: zhartas@gmail.com
        type: string
      name:
        example: zhartas
        type: string
      role:
        example: administrator
        type: string
    type: object
info:
  contact: {}
paths:
  /project:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Object'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Get all projects
      tags:
      - project
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Create a project
      tags:
      - project
  /project/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Delete project by id
      tags:
      - project
    get:
      consumes:
      - application/json
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Get project by id
      tags:
      - project
    put:
      consumes:
      - application/json
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Update project by id
      tags:
      - project
  /project/{id}/tasks:
    get:
      consumes:
      - application/json
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Object'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Get tasks by project id
      tags:
      - project
  /project/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: title
        in: query
        name: title
        type: string
      - description: manager_id
        in: query
        name: manager_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Object'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Search project by param
      tags:
      - project
  /task:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.Object'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: List tasks
      tags:
      - task
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Create a task
      tags:
      - task
  /task/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Delete task by id
      tags:
      - task
    get:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Get task by id
      tags:
      - task
    put:
      consumes:
      - application/json
      parameters:
      - description: task id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/task.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Update task by id
      tags:
      - task
  /task/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: title
        in: query
        name: title
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: priority
        in: query
        name: priority
        type: string
      - description: responsible_id
        in: query
        name: responsibleID
        type: string
      - description: project_id
        in: query
        name: projectID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: Search tasks
      tags:
      - task
  /user:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
      summary: get all users
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
      summary: create new user
      tags:
      - user
  /user/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
      summary: delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
      summary: get user by id
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.Request'
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
      summary: update user data
      tags:
      - user
  /user/{id}/tasks:
    get:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: user tasks by id
      tags:
      - user
  /user/search:
    get:
      consumes:
      - application/json
      description: fill only one and leave second empty
      parameters:
      - description: Optional* Name
        in: query
        name: name
        type: string
      - description: Optional* Email
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Object'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Object'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Object'
      summary: search by name or email
      tags:
      - user
swagger: "2.0"